Revert "Added a simple presentation layer and renamed some variables"
