+++ ClassLibrary/clsBookStock.cs
@@ -2,11 +2,11 @@
 
 namespace ClassLibrary
 {
-    public class clsStock
+    public class clsBookStock
     {
         public bool BookAvailability { get; set; }
         public DateTime DateReceived { get; set; }
-        public int BookId { get; set; }
+        public int BookID { get; set; }
         public string BookTitle { get; set; }
         public int BookQuantity { get; set; }
         public double BookPrice { get; set; }
