+++ Testing1/TestingStock.csproj
@@ -50,7 +50,7 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="tstStock.cs" />
+    <Compile Include="tstBookStock.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
